# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/orders/": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "Gets the list of all available orders.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/orders/{orderId}": {
      "put": {
        "tags": [
          "Orders"
        ],
        "description": "Edits an existing order.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/orders/{orderId}/": {
      "delete": {
        "tags": [
          "Orders"
        ],
        "description": "Deletes a order.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
...
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Gets the list of all available Categories.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/categories/add": {
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "Creates a new category.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/categories/edit/{categoryId}": {
      "put": {
        "tags": [
          "Categories"
        ],
        "description": "Edits an existing category.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/categories/delete/{categoryId}": {
      "delete": {
        "tags": [
          "Categories"
        ],
        "description": "Deletes a category.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/memberships/": {
      "get": {
        "tags": [
          "Memberships"
        ],
        "description": "Gets the list of all available memberships.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/memberships/add": {
      "post": {
        "tags": [
          "Memberships"
        ],
        "description": "Creates a new membership.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "example": "any"
                },
                "discount": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/memberships/edit/{membershipId}": {
      "put": {
        "tags": [
          "Memberships"
        ],
        "description": "Edits an existing membership.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "membershipId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "example": "any"
                },
                "discount": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/memberships/delete/{membershipId}": {
      "delete": {
        "tags": [
          "Memberships"
        ],
        "description": "Deleting an existing membership.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "membershipId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    }
  },
  "definitions": {
...
    "Category": {
      "type": "object",
      "properties": {
        "Name": {
          "type": "string",
          "example": "Watch"
        }
      },
      "required": [
        "Name"
      ]
    },
...
  }
}